HW 2 - Implemented game logic and state.
Added/changed some test implementation, fixed some bugs, for the below:

Added JUMP action to simplify chain moves.  For example, if you have a
roll of 6, and choose to MOVE (6 spaces) --> STACK --> MOVE(jump) -->
TAKE_SHORTCUT --> MOVE/TAXI, it would be hard to keep track of dice state
and differentiate between a regular MOVE (based on die) and JUMP (based on
where you landed) (what if you rolled a 4?).

Changed implementation slightly - now player checks if there is
jump/shortcut/etc and does not roll the die (if necessary) until
afterwards.  Also the TAKE_SHORTCUT was slightly modified such that a
player can signify deciding not to take the shortcut by using this action
with no changes to the pieces in the state.