HW3 - Adding AeroplaneChessPresenter and AeroplaneChessPresenterTest
Also moved common constants to a Constants class (mostly strings and ints
that are used in Logic, Presenter, and any testing), optimized some code
and fixed some bugs in AeroplaneChessLogic.  Changed the implementation of
getOperationsMove, getOperationsTaxi, etc. to fit easily with the
Presenter (now uses just the list of pieces moved, previously used a list
of all the operations).