HW4 - Graphics in GWT.
URL for game: http://1-dot-aeroplane-chess.appspot.com

Made a few changes to AeroplaneCHessPresenter (and corresponding changes
to the Test) to make it easier to implement the UI - ie., when choosing
pieces to move, the player can send any ONE piece on that space (since
some may be stacked/located on the same space and the player will click
only one) and the Presenter will check to see if it is stacked, if so it
will get the other pieces to be moved.  This way the UI does not have to
handle the logic of getting all the stacked pieces.  Also changed the
Presenter so that the viewer and other player won't get the messages if it
isn't their turn (originally going to display the die and moves for the
other player/viewer but since there are no animations yet I just display
the die for the viewer), changed logic tests to match.  Another change
made to the Logic to make UI handling easier was to change representation
of the Launch zone (since there is the same physical location on the board
for all the player's pieces).  Finally, I changed dieRolled and
choosePieces in the Presenter and added some more methods the View can
call such that the View can display a messag eif there is an automatic
jump or "back to the hangar" or other such move.  Fixed some bugs in the
logic (changed to hard-coding some values where the abs value/offset
wasn't working quite correctly in calculating where some areas on the
board start/end).